eliminar task completado
