+++ front-end/js-web-dom/index.html
@@ -17,7 +17,7 @@
     />
     <link rel="stylesheet" href="styles.css" />
   </head>
-  
+
   <body>
     <main>
       <div class="mainCard">
@@ -41,6 +41,6 @@
         </ul>
       </div>
     </main>
-    <script src="script.js"></script>
+    <script type="module" src="script.js"></script>
   </body>
 </html>
