+++ front-end/js-web-dom/script.js
@@ -1,4 +1,6 @@
-(() => {
+  import checkComplete from "./components/checkComplete.js";
+  import deleteIcon from "./components/deleteIcon.js";
+  
   const btn = document.querySelector('[data-form-btn]');
 
   const createTask = (evento) => {
@@ -12,35 +14,17 @@
     
     //backticks
     const taskContent = document.createElement('div');
-
     const titleTask = document.createElement('span');
     titleTask.classList.add('task');
     titleTask.innerText = value;
     taskContent.appendChild(checkComplete());
     taskContent.appendChild(titleTask);
-
-    const content = `
-    <i class="fas fa-trash-alt trashIcon icon"></i>`;
+    
     // task.innerHTML = content;
     task.appendChild(taskContent);
+    task.appendChild(deleteIcon());
     list.appendChild(task);
   };
 
   //Arrow functions o funciones anonimas
   btn.addEventListener('click', createTask);
-
-  const checkComplete = () => {
-    const i = document.createElement('i');
-    i.classList.add('far', 'fa-check-square', 'icon');
-    i.addEventListener('click', completeTask);
-    return i;
-  };
-
-  // Immediately invoked function expression IIFE
-  const completeTask = (event) => {
-    const element = event.target;
-    element.classList.toggle('fas');
-    element.classList.toggle('completeIcon');
-    element.classList.toggle('far');
-  };
-})();
